Renamed PingWebHook* to WebHook\Ping*
